requirements.txt
torch
torchvision
pandas
scikit-learn
numpy
streamlit
Pillow
matplotlib
seaborn

